Made the Z field Integer, so NULL is supported
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/cdk/trunk@10959 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/test/org/openscience/cdk/io/CrystClustReaderTest.java
@@ -83,7 +83,7 @@ public class CrystClustReaderTest extends CDKTestCase {
         org.openscience.cdk.interfaces.ICrystal crystal = model.getCrystal();
         assertNotNull(crystal);
         assertEquals(42, crystal.getAtomCount());
-        assertEquals(1, crystal.getZ());
+        assertEquals(1, crystal.getZ().intValue());
 
         // test reading of partial charges
         org.openscience.cdk.interfaces.IAtom atom = crystal.getAtom(0);
